MLOps_project/requirements.txt
torch
numpy
pandas
torchvision
pillow
typer
pytest
timm
datetime
omegaconf
hydra-core>=1.3
logging
